+++ src/adapters/sqlite/__test__/store.test.ts
@@ -1,0 +1,7 @@
+import { describeAnalysisStore } from '../../../ports/__test__/storageContract'
+import { createDatabase } from '../database'
+import { createSqliteStore } from '../store'
+
+describeAnalysisStore('SQLite in memory', () =>
+  createSqliteStore(createDatabase()),
+)