+++ src/services/__test__/GitService.headBlame.test.ts
@@ -6,8 +6,8 @@ import {
   createTestRepo,
   type TestRepo,
 } from '../../__test__/helpers/createTestRepo'
-import { createDatabase } from '../../db/database'
-import { authors, blameLines, files } from '../../db/schema'
+import { createDatabase } from '../../adapters/sqlite/database'
+import { authors, blameLines, files } from '../../adapters/sqlite/schema'
 import { GitService } from '../GitService'
 
 const GORVEK = { name: 'Gorvek the Ironbane', email: 'gorvek@ashendale.realm' }